+++ src/sensors/agm_icm20948.cpp
@@ -265,11 +265,11 @@ uint32_t AccelIcm20948::SamplingFrequency(uint32_t Freq)
 {
 	// ODR = 1.125 kHz/(1+ACCEL_SMPLRT_DIV[11:0])
 
-	uint32_t div = (1125000 + (Freq >> 1))/ Freq - 1;
+	uint16_t div = (1125000 + (Freq >> 1))/ Freq - 1;
 	div = EndianCvt16(div);
 
 	uint16_t regaddr = ICM20948_ACCEL_SMPLRT_DIV_1_REG;
-	Write16((uint8_t*)&regaddr, 2, div);
+	Write16((uint8_t*)&regaddr, 2, (uint16_t)div);
 
 	return AccelSensor::SamplingFrequency(1125000 / (1 + div));
 }
